Move fetchall_as_namedtuple to avoid import cycle
